+++ packages/date-picker/src/utils/stateful/config/config.impl.ts
@@ -4,11 +4,11 @@ import {
   DEFAULT_LOCALE,
 } from '@schedule-x/shared/src/values'
 import { WeekDay } from '@schedule-x/shared/src/enums/time/week-day.enum'
-import { toDateString } from '@schedule-x/shared/src/utils/stateless/time/format-conversion/date-to-strings'
 import { Placement } from '@schedule-x/shared/src/interfaces/date-picker/placement.enum'
 import { DatePickerListeners } from '@schedule-x/shared/src/interfaces/date-picker/listeners.interface'
 import { DatePickerStyle } from '@schedule-x/shared/src/interfaces/date-picker/style.interface'
 import { signal, Signal } from '@preact/signals'
+import { toDateString } from '@schedule-x/shared/src'
 
 export class ConfigImpl implements DatePickerConfigInternal {
   locale: Signal<string>
@@ -19,7 +19,7 @@ export class ConfigImpl implements DatePickerConfigInternal {
     firstDayOfWeek: WeekDay = DEFAULT_FIRST_DAY_OF_WEEK,
     public min: string = toDateString(new Date(1970, 0, 1)),
     public max: string = toDateString(
-      new Date(new Date().getFullYear() + 1, 11, 31)
+      new Date(new Date().getFullYear() + 50, 11, 31)
     ),
     public placement: Placement = Placement.BOTTOM_START,
     public listeners: DatePickerListeners = {},