+++ packages/translations/src/locales/sr-RS/date-picker.ts
@@ -1,10 +1,9 @@
 import { DatePickerTranslations } from '../../types/date-picker.translations'
 
-export const datePickerSrRs: DatePickerTranslations = {
+export const datePickerSrRS: DatePickerTranslations = {
   Date: 'Datum',
   'MM/DD/YYYY': 'DD/MM/YYYY',
   'Next month': 'Sledeći mesec',
   'Previous month': 'Prethodni mesec',
   'Choose Date': 'Izaberite datum',
-  'Select View': 'Odaberite pregled',
 }