+++ packages/shared/CHANGELOG.md
@@ -3,6 +3,20 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.4.3](https://github.com/schedule-x/schedule-x/compare/v2.4.2...v2.4.3) (2024-11-07)
+
+### Bug Fixes
+
+- missing \_options after dnd ([#761](https://github.com/schedule-x/schedule-x/issues/761)) ([ebe80a6](https://github.com/schedule-x/schedule-x/commit/ebe80a6ce3de5f834ddea6e35090aac661b0ccc2))
+
+## [2.4.2](https://github.com/schedule-x/schedule-x/compare/v2.4.1...v2.4.2) (2024-11-06)
+
+**Note:** Version bump only for package @schedule-x/shared
+
+## [2.4.1](https://github.com/schedule-x/schedule-x/compare/v2.4.0...v2.4.1) (2024-10-10)
+
+**Note:** Version bump only for package @schedule-x/shared
+
 # [2.4.0](https://github.com/schedule-x/schedule-x/compare/v2.3.0...v2.4.0) (2024-10-04)
 
 **Note:** Version bump only for package @schedule-x/shared